+++ Runtime/SceneManager/SceneManager.cs
@@ -13,17 +13,39 @@ namespace ActionCode.SceneManagement
     /// The Scene Manager. It loads new Scenes by first opening a LoadingScene and showing fade transitions.
     /// </summary>
     [CreateAssetMenu(fileName = "SceneManager", menuName = "ActionCode/SceneManager/Scene Manager", order = 110)]
-    public sealed class SceneManager : ScriptableObject, ISceneManager
+    public sealed class SceneManager : ScriptableObject
     {
         [CreateButton(typeof(SceneTransition))]
         [Tooltip("De default Scene Transition values used when none is provided.")]
         public SceneTransition defaultTransition;
 
+        /// <summary>
+        /// Action fired when loading is started.
+        /// </summary>
         public event Action OnLoadingStarted;
+
+        /// <summary>
+        /// Action fired when loading is finished.
+        /// </summary>
         public event Action OnLoadingFinished;
+
+        /// <summary>
+        /// Action fired when the loading progress changes.
+        /// <para>
+        /// The given float param is the progress in percentage, 
+        /// i.e. a number between 0F and 100F.
+        /// </para>
+        /// </summary>
         public event Action<float> OnProgressChanged;
 
+        /// <summary>
+        /// Whether a Scene loading process is happening. 
+        /// </summary>
         public bool IsLoading { get; private set; }
+
+        /// <summary>
+        /// Whether the Scene loading process is locked. 
+        /// </summary>
         public bool IsLoadingLocked { get; private set; }
 
         private void OnDisable()
@@ -32,15 +54,38 @@ namespace ActionCode.SceneManagement
             UnlockLoading();
         }
 
+        /// <summary>
+        /// Loads the given Scene while showing a LoadingScene and fade 
+        /// transitions using the given transition data.
+        /// </summary>
+        /// <param name="scene">The name or path of the Scene to load.</param>
+        /// <param name="transition">The transition data to use.</param>
         public void LoadScene(string scene, SceneTransition transition = null) =>
             _ = LoadSceneAsync(scene, transition != null ? transition : defaultTransition);
 
         public async Task LoadSceneAsync(string scene) => await LoadSceneAsync(scene, defaultTransition);
 
+        /// <summary>
+        /// Loads the given Scene asynchronously while showing a LoadingScene and fade 
+        /// transitions using the given transition data.
+        /// </summary>
+        /// <param name="scene"><inheritdoc cref="LoadSceneAsync(string)"/></param>
+        /// <param name="transition">The transition data to use.</param>
+        /// <returns><inheritdoc cref="LoadSceneAsync(string)"/></returns>
         public async Task LoadSceneAsync(string scene, SceneTransition transition) =>
             await AwaitableCoroutine.Run(LoadSceneCoroutine(scene, transition));
 
+        /// <summary>
+        /// Locks the Scene loading process.
+        /// <para>
+        /// Use it to lock into the Loading Scene until the <see cref="UnlockLoading"/> is called.
+        /// </para>
+        /// </summary>
         public void LockLoading() => IsLoadingLocked = true;
+
+        /// <summary>
+        /// Unlocks the Scene loading process.
+        /// </summary>
         public void UnlockLoading() => IsLoadingLocked = false;
 
         private IEnumerator LoadSceneCoroutine(string scene, SceneTransition transition)