+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Removed
 - ISceneManager interface
+- Scriptable Settings Provider package dependency
 
 ## [3.1.0] - 2024-10-06
 ### Added