+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [4.1.0] - 2024-12-07
 ### Added
 - SceneManager.QuitGame function
 - SceneManager.LoadScene functions using Scene
@@ -97,7 +99,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - README
 - Initial commit
 
-[Unreleased]: https://github.com/HyagoOliveira/SceneManagement/compare/4.0.0...main
+[Unreleased]: https://github.com/HyagoOliveira/SceneManagement/compare/4.1.0...main
+[4.1.0]: https://github.com/HyagoOliveira/SceneManagement/tree/4.1.0/
 [4.0.0]: https://github.com/HyagoOliveira/SceneManagement/tree/4.0.0/
 [3.1.0]: https://github.com/HyagoOliveira/SceneManagement/tree/3.1.0/
 [3.0.0]: https://github.com/HyagoOliveira/SceneManagement/tree/3.0.0/